pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mmix"
version = "1.0.0"
description = "Production-ready Agentic Marketing Mix Modeling Pipeline"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MMIX Team"}
]
keywords = ["marketing-mix-modeling", "mmm", "optimization", "llm", "agentic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Business/Independent Software Vendors",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "statsmodels>=0.13.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "python-dotenv>=0.19.0",
    "openai>=1.0.0",
    "openpyxl>=3.0.0",
    "python-pptx>=0.6.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.990",
    "isort>=5.10.0",
]

[project.urls]
Homepage = "https://github.com/kunalbhargava/Hackathon"
Documentation = "https://github.com/kunalbhargava/Hackathon/tree/main/docs"
Repository = "https://github.com/kunalbhargava/Hackathon.git"

[tool.setuptools]
packages = ["mmix"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
mmix = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --strict-markers"
